requirements.txt
tensorflow
keras
opencv-python
pillow
pytesseract
